ARM: dts: stm32mp157c-byteengine/bytedevkit: fix qspi_mm region
Using the default value from the stm32mp151.dtsi gives error in the
stm32-qspi probe function on a stm32mp151a with 1 GB RAM:

[    3.194620] vmap allocation for size 268439552 failed:
               use vmalloc=<size> to increase size
[    3.201610] stm32-qspi 58003000.spi: ioremap failed for
               resource [mem 0x70000000-0x7ffffffe]
[    3.210126] stm32-qspi: probe of 58003000.spi failed with error -12

Therefore no mtd device is present in Linux. To avoid that do not map
the full 256 MB address space.
Because only a 16 MB SPI NOR is attached, this should not be an issue.

Signed-off-by: Rico Ganahl <rico.ganahl@bytesatwork.ch>